% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_cities.R
\name{get_all_cities}
\alias{get_all_cities}
\title{Get vector of all cities}
\usage{
get_all_cities()
}
\value{
Vector with all cities as string
}
\description{
Get a vector of all possible cities, that are covered by the KPI-information
}
\examples{
get_all_cities()

}
