% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_kpis.R
\name{get_all_kpis}
\alias{get_all_kpis}
\title{Get overview of all possible KPIs}
\usage{
get_all_kpis()
}
\value{
Dataframe with KPI-Id, KPI-title, group-Id and title it belongs to
}
\description{
Get all KPI-groups with ids and title together with their individual KPI members including
member_title and member_id.
}
\examples{
get_all_kpis()

}
