% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infos.R
\name{get_infos}
\alias{get_infos}
\title{Request data from Kolade.se}
\usage{
get_infos(kpis, cities, years)
}
\arguments{
\item{kpis}{Vector of requested KPIs (string format)}

\item{cities}{Vector of requested cities (string format}

\item{years}{Vector of requested years (numbers)}
}
\value{
Dataframe that contains the requested KPI data
}
\description{
Request KPI data from Kolada.se for multiple cities and years
}
\examples{
get_infos(kpis = c("N02904", "U28116"), cities = c("Helsingborg", "Askersund"), years = c(2019, 2013))

}
